requirements.txt
PyQt6>=6.6.0
PyMuPDF>=1.24.0
openai>=1.30.0
anthropic>=0.49.0
openpyxl>=3.1.0
Pillow>=10.0.0
python-docx>=1.0.0
